Restructure the Flask app:
Tried to untangle the cyclic dependencies, although there are still some of
them left. The main reason for this restructuring/refactoring is to make
the application a bit more extensible, so it's possible to add more routes
to it without touching the core.
