+++ redash/__init__.py
@@ -1,11 +1,16 @@
 import logging
 import urlparse
 import redis
+from flask import Flask
+from flask_sslify import SSLify
+from werkzeug.contrib.fixers import ProxyFix
+from werkzeug.routing import BaseConverter, ValidationError
 from statsd import StatsClient
 from flask_mail import Mail
 
 from redash import settings
 from redash.query_runner import import_query_runners
+
 
 __version__ = '0.10.0'
 
@@ -49,3 +54,54 @@ import_query_runners(settings.QUERY_RUNNERS)
 
 from redash.version_check import reset_new_version_status
 reset_new_version_status()
+
+
+class SlugConverter(BaseConverter):
+    def to_python(self, value):
+        # This is an ugly workaround for when we enable multi-org and some files are being called by the index rule:
+        if value in ('google_login.png', 'favicon.ico', 'robots.txt', 'views'):
+            raise ValidationError()
+
+        return value
+
+    def to_url(self, value):
+        return value
+
+
+def create_app():
+    from redash import handlers
+    from redash.admin import init_admin
+    from redash.models import db
+    from redash.authentication import setup_authentication
+    from redash.metrics.request import provision_app
+
+    app = Flask(__name__,
+                template_folder=settings.STATIC_ASSETS_PATH,
+                static_folder=settings.STATIC_ASSETS_PATH,
+                static_path='/static')
+
+    # Make sure we get the right referral address even behind proxies like nginx.
+    app.wsgi_app = ProxyFix(app.wsgi_app, settings.PROXIES_COUNT)
+    app.url_map.converters['org_slug'] = SlugConverter
+
+    if settings.ENFORCE_HTTPS:
+        SSLify(app, skips=['ping'])
+
+    if settings.SENTRY_DSN:
+        from raven.contrib.flask import Sentry
+        sentry = Sentry(app, dsn=settings.SENTRY_DSN)
+        sentry.client.release = __version__
+
+    # configure our database
+    settings.DATABASE_CONFIG.update({'threadlocals': True})
+    app.config['DATABASE'] = settings.DATABASE_CONFIG
+    app.config.update(settings.all_settings())
+
+    provision_app(app)
+    init_admin(app)
+    db.init_app(app)
+    mail.init_app(app)
+    setup_authentication(app)
+    handlers.init_app(app)
+
+    return app
