+++ redash/authentication/google_oauth.py
@@ -79,7 +79,7 @@ def org_login(org_slug):
 @blueprint.route('/oauth/google', endpoint="authorize")
 def login():
     callback = url_for('.callback', _external=True)
-    next = request.args.get('next', url_for("index", org_slug=session.get('org_slug')))
+    next = request.args.get('next', url_for("redash.index", org_slug=session.get('org_slug')))
     logger.debug("Callback url: %s", callback)
     logger.debug("Next is: %s", next)
     return google_remote_app().authorize(callback=callback, state=next)
@@ -93,12 +93,12 @@ def authorized():
     if access_token is None:
         logger.warning("Access token missing in call back request.")
         flash("Validation error. Please retry.")
-        return redirect(url_for('login'))
+        return redirect(url_for('redash.login'))
 
     profile = get_user_profile(access_token)
     if profile is None:
         flash("Validation error. Please retry.")
-        return redirect(url_for('login'))
+        return redirect(url_for('redash.login'))
 
     if 'org_slug' in session:
         org = models.Organization.get_by_slug(session.pop('org_slug'))
@@ -108,10 +108,10 @@ def authorized():
     if not verify_profile(org, profile):
         logger.warning("User tried to login with unauthorized domain name: %s (org: %s)", profile['email'], org)
         flash("Your Google Apps domain name isn't allowed.")
-        return redirect(url_for('login', org_slug=org.slug))
+        return redirect(url_for('redash.login', org_slug=org.slug))
 
     create_and_login_user(org, profile['name'], profile['email'])
 
-    next = request.args.get('state') or url_for("index", org_slug=org.slug)
+    next = request.args.get('state') or url_for("redash.index", org_slug=org.slug)
 
     return redirect(next)
