+++ redash/authentication/org_resolving.py
@@ -6,9 +6,11 @@ single Organization in your installation.
 """
 
 import logging
-from redash.models import Organization
-from werkzeug.local import LocalProxy
+
 from flask import request
+from werkzeug.local import LocalProxy
+
+from redash.models import Organization
 
 
 def _get_current_org():
@@ -16,7 +18,6 @@ def _get_current_org():
     org = Organization.get_by_slug(slug)
     logging.debug("Current organization: %s (slug: %s)", org, slug)
     return org
-
 
 # TODO: move to authentication
 current_org = LocalProxy(_get_current_org)
