+++ redash/handlers/groups.py
@@ -2,7 +2,6 @@ import time
 from flask import request
 from flask_restful import abort
 from redash import models
-from redash.wsgi import api
 from redash.permissions import require_admin, require_permission
 from redash.handlers.base import BaseResource, get_object_or_404
 
@@ -179,9 +178,3 @@ class GroupDataSourceResource(BaseResource):
         })
 
 
-api.add_org_resource(GroupListResource, '/api/groups', endpoint='groups')
-api.add_org_resource(GroupResource, '/api/groups/<group_id>', endpoint='group')
-api.add_org_resource(GroupMemberListResource, '/api/groups/<group_id>/members', endpoint='group_members')
-api.add_org_resource(GroupMemberResource, '/api/groups/<group_id>/members/<user_id>', endpoint='group_member')
-api.add_org_resource(GroupDataSourceListResource, '/api/groups/<group_id>/data_sources', endpoint='group_data_sources')
-api.add_org_resource(GroupDataSourceResource, '/api/groups/<group_id>/data_sources/<data_source_id>', endpoint='group_data_source')
